# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateEventInquiryRequest {
  /** @minLength 1 */
  full_name: string;
  /** @minLength 1 */
  phone: string;
  telegram_username?: string;
  /** @minLength 1 */
  event_type: string;
  event_date: string;
  /** @minimum 1 */
  guest_count: number;
  notes?: string;
}

export type EventInquiryStatus = typeof EventInquiryStatus[keyof typeof EventInquiryStatus];


export const EventInquiryStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export interface EventInquiry {
  id: string;
  created_at: string;
  full_name: string;
  phone: string;
  telegram_username?: string;
  event_type: string;
  event_date: string;
  guest_count: number;
  notes?: string;
  status: EventInquiryStatus;
}

export interface ValidationError {
  title: string;
  detail: string;
}