# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Private event inquiry operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /events/inquire:
    post:
      operationId: createEventInquiry
      tags: [events]
      summary: Submit a private event inquiry
      description: Creates a new private event booking inquiry and notifies the owner via Telegram.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInquiryRequest"
      responses:
        "201":
          description: Inquiry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventInquiry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateEventInquiryRequest:
      type: object
      required:
        - full_name
        - phone
        - event_type
        - event_date
        - guest_count
      properties:
        full_name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        telegram_username:
          type: string
        event_type:
          type: string
          minLength: 1
        event_date:
          type: string
          format: date
        guest_count:
          type: integer
          minimum: 1
        notes:
          type: string
    EventInquiry:
      type: object
      required:
        - id
        - created_at
        - full_name
        - phone
        - event_type
        - event_date
        - guest_count
        - status
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        full_name:
          type: string
        phone:
          type: string
        telegram_username:
          type: string
        event_type:
          type: string
        event_date:
          type: string
        guest_count:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, declined]
    ValidationError:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
      required:
        - title
        - detail