# lib/api-zod/src/generated/types/createEventInquiryRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventInquiryRequest {
  /** @minLength 1 */
  full_name: string;
  /** @minLength 1 */
  phone: string;
  telegram_username?: string;
  /** @minLength 1 */
  event_type: string;
  event_date: Date;
  /** @minimum 1 */
  guest_count: number;
  notes?: string;
}