# lib/api-zod/src/generated/types/eventInquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventInquiryStatus } from './eventInquiryStatus';

export interface EventInquiry {
  id: string;
  created_at: Date;
  full_name: string;
  phone: string;
  telegram_username?: string;
  event_type: string;
  event_date: string;
  guest_count: number;
  notes?: string;
  status: EventInquiryStatus;
}